+++ Changelog.md
@@ -9,6 +9,7 @@ Rust:
 - Add function (sub-circuit) into the builder with a FunctionBuilder
 - Format the entire project with rustfmt
 - Add SwitchBuilder
+- When declaring a function, limit the use of Copy gates
 
 # v2.0.0, 2021-12